fix(user): user와 user profile 분리
