+++ src/shared/schemas/user/index.ts
@@ -13,17 +13,6 @@ export const UserEntitySchema = z.strictObject({
 	emailVerified: z.nullable(z.date()),
 })
 
-export const UserProfileEntitySchema = z.strictObject({
-	id: z.string().check(z.cuid()),
-	userId: z.string().check(z.cuid()),
-	createdAt: z.date(),
-	updatedAt: z.date(),
-	deletedAt: z.nullable(z.date()),
-	phone: z.nullable(z.string()),
-	birthDate: z.nullable(z.date()),
-	organization: z.nullable(z.string()),
-})
-
 export const UpdateUserInputSchema = z.object({
 	name: z
 		.string()
@@ -34,41 +23,4 @@ export const UpdateUserInputSchema = z.object({
 	email: z.string().check(z.email('올바른 이메일 형식이 아닙니다')),
 })
 
-export const GetCurrentUserResponseSchema = UserEntitySchema
-
-export const UserProfileStatsSchema = z.object({
-	totalApplications: z.number().check(z.nonnegative()),
-	selectedApplications: z.number().check(z.nonnegative()),
-	completedActivities: z.number().check(z.nonnegative()),
-	memberSince: z.date(),
-})
-
-export const DeleteAccountInputSchema = z.object({
-	confirmEmail: z.string().check(z.email('올바른 이메일 형식이 아닙니다')),
-	confirmText: z.literal('계정을 삭제합니다'),
-})
-
-export const CreateUserProfileInputSchema = z.strictObject({
-	phone: z.optional(
-		z.nullable(z.string().check(z.minLength(0), z.maxLength(11))),
-	),
-	birthDate: z.optional(z.nullable(z.date())),
-	organization: z.optional(
-		z.nullable(
-			z.pipe(
-				z.string().check(z.minLength(0), z.maxLength(255)),
-				z.transform((organization) =>
-					organization.trim() === '' ? null : organization,
-				),
-			),
-		),
-	),
-})
-
-export const UpdateUserProfileInputSchema = CreateUserProfileInputSchema
-
-export const GetUserProfileResponseSchema = UserProfileEntitySchema
-
-export const HasUserProfileResponseSchema = z.strictObject({
-	hasProfile: z.boolean(),
-})
+export const GetUserResponseSchema = UserEntitySchema
